Fixed compile error (missing uin64 cast).
#ROBOMERGE-AUTHOR: ionut.matasaru
#ROBOMERGE-SOURCE: CL 18166438 in //UE5/Release-5.0/...
#ROBOMERGE-BOT: STARSHIP (Release-5.0 -> Release-Engine-Staging) (v894-18169371)

[CL 18176603 by ionut matasaru in ue5-release-engine-staging branch]

+++ Engine/Source/Runtime/Core/Private/ProfilingDebugging/MemoryAllocationTrace.cpp
@@ -280,14 +280,14 @@ void FAllocationTrace::ReallocFree(void* Address, HeapId RootHeap)
 		case EMemoryTraceRootHeap::SystemMemory:
 		{
 			UE_TRACE_LOG(Memory, ReallocFreeSystem, MemAllocChannel)
-				<< ReallocFreeSystem.Address(Address);
+				<< ReallocFreeSystem.Address(uint64(Address));
 			break;
 		}
 
 		default:
 		{
 			UE_TRACE_LOG(Memory, ReallocFree, MemAllocChannel)
-				<< ReallocFree.Address(Address)
+				<< ReallocFree.Address(uint64(Address))
 				<< ReallocFree.RootHeap(uint8(RootHeap));
 			break;
 		}
